Version 3.8.3
getdate: Print hjd in help message instead of jhd
gethead: First column name is FILE_NAME if FILENAME keyword is used (2011-06-29)
gethead: Fix memory leaks (2011-05-19)
gethead: Add -j option to convert RA and Dec to J2000 (2011-05-20)
gethead: Fix -x option for named or unnamed range of extensions (2011-08-12)
gettab: Fix variable name bug so program works!
scat: Fix tab spacing in tmce(2MASS catalog with magnitude errors) (2011-04-21)
imcat,scat,imwcs,immatch: Fix problem with UCAC3 (2011-04-27)
skycoor: Deal with comma in command line input coordinates (2011-04-19)
skycoor: Drop 99 for skipped RA in input list file (2011-08-17)
hget.c: Free allocated memory in strnsrch() to eliminate memory leak
imhfile.c: Free *newpixname* not pixname. (2011-05-20)
ned_cli.c: Declard *d double, not long double, in ned_getdouble() (2011-05-20)
ucacread.c: Find stars when RA is unevenly distributed in a declination zone
ucacread.c: Drop extra formats from printf() (2011-05-20)
wcs.c: Fix wcsfree() so it frees depended-on WCS structures (2011-05-09)
wcsinit.c: Change error() calls to setwcserr()
wcslib.h: Declare undeclared SCAMP subroutine raw-to-pv()

+++ libned/ned_cli.c
@@ -144,7 +144,7 @@ int *n;
 
 int
 ned_getdouble(d)
-long double *d;
+double *d;
 {
    char s[STR_LENGTH+1]; 
 
@@ -310,4 +310,6 @@ double radius;
 }
 
 /* Sep 18 2007	Modified to avoid use of vararg for portability
+ *
+ * May 20 2011	Fixed declaration of *d to double in ned_getdouble()
  */
